feat: реализован функционал аутентификации и логина

+++ controllers/cards.js
@@ -1,4 +1,7 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
+const BadRequestError = require('../errors/bad-req-err');
+const UnauthorizedError = require('../errors/unauthorized-err');
 
 // const checkCard = (err, res) => {
 //   if (err.kind === 'ObjectId') {
@@ -8,39 +11,48 @@ const Card = require('../models/card');
 //   }
 // };
 
+const checkOwner = (cardId, userId) => {
+  Card.findById(cardId).then((card) => {
+    if (!card.owner === userId) {
+      throw new UnauthorizedError('Ошибка аутентификации');
+    }
+  });
+};
+
 const checkDate = (err, res, errorText) => {
   if (err.name === 'ValidationError') {
-    return res.status(400).send({
-      message: errorText,
-    });
+    throw new BadRequestError('Ошибка в данных');
   }
 };
 
 module.exports.getCards = (req, res) => {
   Card.find({})
     .then((card) => res.send({ data: card }))
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
+module.exports.deleteCard = (req, res, next) => {
+  const owner = req.user._id;
+
+  checkOwner(req.params.cardId, owner);
+
   Card.findByIdAndRemove(req.params.cardId)
     .then((card) => {
       if (!card) {
-        return res
-          .status(404)
-          .send({ message: 'Карточка с указанным _id не найдена' });
+        throw new NotFoundError('Карточка не найдена');
       }
       res.send({ data: card });
     })
     .catch((err) => {
       if (err.kind === 'ObjectId') {
-        return res.status(400).send({ message: 'Ошибка в данных' });
+        throw new BadRequestError('Ошибка в данных');
       }
-      res.status(500).send({ message: 'Произошла ошибка' });
-    });
+      next(err);
+    })
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
@@ -48,52 +60,51 @@ module.exports.createCard = (req, res) => {
     .then((card) => res.status(201).send({ data: card }))
     .catch((err) => {
       checkDate(err, res, 'Переданы некорректные данные при создании карточки');
-      res.status(500).send({ message: 'Произошла ошибка' });
-    });
+      next(err);
+    })
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   // eslint-disable-next-line implicit-arrow-linebreak
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
-    { new: true, runValidators: true },
+    { new: true, runValidators: true }
   )
     .then((card) => {
       if (!card) {
-        return res
-          .status(404)
-          .send({ message: 'Карточка с указанным _id не найдена' });
+        throw new NotFoundError('Карточка не найдена');
       }
       res.send({ data: card });
     })
     .catch((err) => {
       if (err.kind === 'ObjectId') {
-        return res.status(400).send({ message: 'Ошибка в данных' });
+        throw new BadRequestError('Ошибка в данных');
       }
-      res.status(500).send({ message: 'Произошла ошибка' });
-    });
+      next(err);
+    })
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   // eslint-disable-next-line implicit-arrow-linebreak
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
-    { new: true, runValidators: true },
+    { new: true, runValidators: true }
   )
     .then((card) => {
       if (!card) {
-        return res
-          .status(404)
-          .send({ message: 'Карточка с указанным _id не найдена' });
+        throw new NotFoundError('Карточка не найдена');
       }
       res.send({ data: card });
     })
     .catch((err) => {
       if (err.kind === 'ObjectId') {
-        return res.status(400).send({ message: 'Ошибка в данных' });
+        throw new BadRequestError('Ошибка в данных');
       }
-      res.status(500).send({ message: 'Произошла ошибка' });
-    });
+      next(err);
+    })
+    .catch(next);
 };
